+++ src/services/bakingBad/index.ts
@@ -1,6 +1,7 @@
-import { Network } from "../beacon/context";
+import { Network } from "services/beacon";
 
 export const API_URL = "https://api.better-call.dev/v1";
+
 export const networkNameMap: Record<Network, string> = {
   mainnet: "mainnet",
   hangzhounet: "hangzhou2net",
