+++ src/services/bakingBad/transfers/index.ts
@@ -1,4 +1,4 @@
-import { Network } from "services/beacon/context";
+import { Network } from "services/beacon";
 import {API_URL, networkNameMap} from "..";
 import { TransferDTO, TransfersDTO } from "./types";
